+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - DH
+### Added
+- Flights model
+
+### Changed
+- Autoload.php to include Flights model
+
 ## [Unreleased] - DH & YR
 ### Added
 - Airlines Model which uses curl to retrieve all the airlines